requirements.txt
celery==5.2.7
Django==4.1.4
django-allauth==0.51.0
django-environ==0.9.0
django-extensions==3.2.1
django-redis==5.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
flake8==6.0.0
isort==5.11.4
Pillow==9.3.0
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-django==4.5.2
redis==4.4.0
stripe==5.0.0
